Refactor contract model and schema to use Decimal for quantity and price fields
- Updated the Contract model to replace float with Decimal for quantity_mwh and price_per_mwh, ensuring better precision.
- Modified the ContractBase and ContractUpdate schemas to reflect these changes, including validation constraints.
- Adjusted the SQL schema to define quantity_mwh and price_per_mwh as NUMERIC types for consistency.

+++ backend/app/models.py
@@ -1,6 +1,7 @@
 from datetime import date
+from decimal import Decimal
 
-from sqlalchemy import Date, Float, Integer, String
+from sqlalchemy import Date, Integer, Numeric, String
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 
@@ -13,8 +14,8 @@ class Contract(Base):
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
     energy_type: Mapped[str] = mapped_column(String(50), nullable=False)
-    quantity_mwh: Mapped[float] = mapped_column(Float, nullable=False)
-    price_per_mwh: Mapped[float] = mapped_column(Float, nullable=False)
+    quantity_mwh: Mapped[Decimal] = mapped_column(Numeric(18, 3), nullable=False)
+    price_per_mwh: Mapped[Decimal] = mapped_column(Numeric(18, 6), nullable=False)
     delivery_start: Mapped[date] = mapped_column(Date, nullable=False)
     delivery_end: Mapped[date] = mapped_column(Date, nullable=False)
     location: Mapped[str] = mapped_column(String(80), nullable=False)
